fix!: refactor propagators to add #fields
