+++ api/lib/opentelemetry/trace/propagation/trace_context.rb
@@ -5,8 +5,7 @@
 # SPDX-License-Identifier: Apache-2.0
 
 require 'opentelemetry/trace/propagation/trace_context/trace_parent'
-require 'opentelemetry/trace/propagation/trace_context/text_map_extractor'
-require 'opentelemetry/trace/propagation/trace_context/text_map_injector'
+require 'opentelemetry/trace/propagation/trace_context/text_map_propagator'
 
 module OpenTelemetry
   module Trace
@@ -15,24 +14,14 @@ module OpenTelemetry
       # for context propagation in the W3C Trace Context format.
       module TraceContext
         extend self
-        TRACEPARENT_KEY = 'traceparent'
-        TRACESTATE_KEY = 'tracestate'
-        TEXT_MAP_EXTRACTOR = TextMapExtractor.new
-        TEXT_MAP_INJECTOR = TextMapInjector.new
+        TEXT_MAP_PROPAGATOR = TextMapPropagator.new
 
-        private_constant :TRACEPARENT_KEY, :TRACESTATE_KEY,
-                         :TEXT_MAP_INJECTOR, :TEXT_MAP_EXTRACTOR
+        private_constant :TEXT_MAP_PROPAGATOR
 
-        # Returns an extractor that extracts context using the W3C Trace Context
-        # format
-        def text_map_extractor
-          TEXT_MAP_EXTRACTOR
-        end
-
-        # Returns an injector that injects context using the W3C Trace Context
-        # format
-        def text_map_injector
-          TEXT_MAP_INJECTOR
+        # Returns a text map propagator that propagates context using the
+        # W3C Trace Context format.
+        def text_map_propagator
+          TEXT_MAP_PROPAGATOR
         end
       end
     end
